Refactor UI components: Extract variants, update imports, and simplify code
